create MyRooms component, sending props from ChatNew + Available Rooms in progress
